POM: switch to jython-standalone JAR
Both jython and jython-standalone are (unfortunately) uber-JARs with
unshaded dependencies. The difference is that jython-standalone also has
the crucial Lib/ directory for core Python packages, as well as quite a
few packageless classes not found in the stock jython JAR, which are
presumably necessary in certain circumstances.

While ideally we would depend on a granular jython JAR file that does
not have _any_ embedded dependencies, for the time being, to fix
problems with Jython scripts that want to use core Python packages, we
need to depend on jython-standalone instead.

Hopefully, this changes addresses Fiji bug 890:
   http://fiji.sc/bugzilla/show_bug.cgi?id=890