JythonScriptLanguage: decode more PyObject types
It seems that for PyInteger at least, even though it is a PyObject, the
approach of calling __tojava__(getType().getProxyType()) does not work.

So to be safe, let's hardcode the types we _know_ we can support --
PyBoolean, PyInteger, PyFloat and PyString -- before PyObject generally.

Hopefully, this will prevent warnings such as:

    [WARNING] Ignoring unsupported output: result [org.python.core.PyInteger]

This change also updates the JythonTest#testBasic() method,
since the result is now an Integer rather than a PyInteger.