JythonTest: decode variables before checking them
I could have sworn that I tested this, but apparently not.

It turns out that return values are not decoded by default.
So for the time being, let's just do it in the test here.
(Ultimately, we probably want to decode them automatically.)

Values extracted directly from the bindings are also not decoded,
 so let's decode those manually as well.
(That should probably also be automatic, though I'm not sure.)